fix: Merge 시 날려버린 코드 복구함

+++ src/main/java/kr/co/fastcampus/travel/domain/itinerary/service/ItineraryService.java
@@ -39,8 +39,12 @@ public class ItineraryService {
     }
 
     @Transactional
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String email) {
         var itinerary = findById(id);
+        String writerEmail = itinerary.getTrip().getMember().getEmail();
+        if (!writerEmail.equals(email)) {
+            throw new MemberMismatchException();
+        }
         itineraryRepository.delete(itinerary);
     }
 }
